Allow thresholding transtion maps
